Move/change New project modal and button
